Clean realtime example code
Delete commented-out code

Improve console output

Extract main routine into smaller functions
