+++ src/sydevs/systems/real_time_buffer.cpp
@@ -76,62 +76,6 @@ void real_time_buffer::recompute_planned_clock_duration()
             syn_count_ = 0;
         }
     }
-
-    /*
-    planned_clock_t_ = clock_time();
-    int64 depth = std::min(ta_depth_, max_depth_);
-    if (depth > 0) {
-        float64 accel_clock_dt_sum = 0.0;
-        float64 decel_clock_dt_sum = 0.0;
-        int64 accel_clock_dt_count = 0;
-        int64 decel_clock_dt_count = 0;
-        const auto& retained_t0 = time_points_[0];
-        const auto& retained_clock_t0 = clock_times_[0];
-        float64 clock_dt0 = 0.0;
-        for (int64 i = 0; i < depth; ++i) {
-            const auto& retained_ti = time_points_[i];
-            const auto& retained_clock_ti = clock_times_[i];
-            auto dti = retained_t0.gap(retained_ti) + planned_dt_.unfixed();
-            auto clock_ti = retained_clock_ti + std::chrono::microseconds(int64((dti/ta_rate_)/1_us));
-            float64 clock_dti = 0.0;
-            if (clock_ti > retained_clock_t0) {
-                clock_dti = float64(std::chrono::duration_cast<std::chrono::microseconds>(clock_ti - retained_clock_t0).count());
-            }
-            if (i == 0) {
-                clock_dt0 = clock_dti;
-            }
-            else {
-                clock_dti = std::min(clock_dti, 2.0*clock_dt0);
-            }
-            if (clock_dti <= clock_dt0) {
-                accel_clock_dt_sum += clock_dti;
-                ++accel_clock_dt_count;
-            }
-            if (clock_dti >= clock_dt0) {
-                decel_clock_dt_sum += clock_dti;
-                ++decel_clock_dt_count;
-            }
-            if (i == depth - 1) {
-                if (clock_dti < clock_dt0) {
-                    decel_clock_dt_sum = 0.0;
-                    decel_clock_dt_count = 0;
-                }
-                else if (clock_dti > clock_dt0) {
-                    accel_clock_dt_sum = 0.0;
-                    accel_clock_dt_count = 0;
-                }
-                else {
-                    accel_clock_dt_sum = clock_dti;
-                    accel_clock_dt_count = 1;
-                    decel_clock_dt_sum = 0.0;
-                    decel_clock_dt_count = 0;
-                }
-            }
-        }
-        float64 clock_dt = (accel_clock_dt_sum + decel_clock_dt_sum)/(accel_clock_dt_count + decel_clock_dt_count);
-        planned_clock_t_ = retained_clock_t0 + std::chrono::microseconds(int64(clock_dt));
-    }
-    */
 }
